+++ test/performance/perf_test.lua
@@ -24,7 +24,7 @@ end
 
 local vshard_cfg_template = {
     sharding = {
-        {
+        ['s-1'] = {
             replicas = {
                 ['s1-master'] = {
                     master = true,
@@ -32,7 +32,7 @@ local vshard_cfg_template = {
                 ['s1-replica'] = {},
             },
         },
-        {
+        ['s-2'] = {
             replicas = {
                 ['s2-master'] = {
                     master = true,
@@ -40,7 +40,7 @@ local vshard_cfg_template = {
                 ['s2-replica'] = {},
             },
         },
-        {
+        ['s-3'] = {
             replicas = {
                 ['s3-master'] = {
                     master = true,